36593 Updated the PR with some refactoring as per the review
